sync-script/requirements.txt
boto3
azure-storage-blob
python-dotenv

